Relocating Timer module classes
