+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/game/Game.java
@@ -37,9 +37,9 @@ import de.butzlabben.missilewars.game.schematics.objects.Missile;
 import de.butzlabben.missilewars.game.schematics.objects.Shield;
 import de.butzlabben.missilewars.game.signs.MWSign;
 import de.butzlabben.missilewars.game.stats.FightStats;
-import de.butzlabben.missilewars.game.timer.EndTimer;
-import de.butzlabben.missilewars.game.timer.GameTimer;
-import de.butzlabben.missilewars.game.timer.LobbyTimer;
+import de.butzlabben.missilewars.game.timer.modules.EndTimer;
+import de.butzlabben.missilewars.game.timer.modules.GameTimer;
+import de.butzlabben.missilewars.game.timer.modules.LobbyTimer;
 import de.butzlabben.missilewars.game.timer.TaskManager;
 import de.butzlabben.missilewars.listener.game.EndListener;
 import de.butzlabben.missilewars.listener.game.GameBoundListener;
